refactor: AppContent를 src 하위로 이동 (#260)
