+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import { BrowserRouter as Router } from 'react-router-dom'
 import './App.css'
-import { AppContent } from './components/layout/AppContent'
+import { BrowserRouter as Router } from 'react-router-dom'
+import { AppContent } from './AppContent'
 
 const App: React.FC = () => {
   return (
